presentML/multi_app_project/requirements.txt
streamlit
pillow
pytesseract
opencv-python
numpy
matplotlib
seaborn
scikit-learn
pandas
plotly
nltk
wordcloud
networkx
google-generativeai
python-dotenv
requests
